Fix some warnings for x64:   dump1090.c(4247,70): warning: cast to 'void *' from smaller integer type 'int' [-Wint-to-void-pointer-cast]            conn = mg_http_listen (&Modes.mgr, url, connection_handler, (void*) service);                                                                        ^~~~~~~~~~~~~~~
Hence use a 'intptr_t' for all 'service' variables.

+++ misc.h
@@ -131,7 +131,7 @@ typedef void (*msg_handler) (mg_iobuf *msg, int loop_cnt);
  */
 typedef struct connection {
         mg_connection     *conn;              /**< Remember which connection this client/server is in */
-        int                service;           /**< This client's service membership */
+        intptr_t           service;           /**< This client's service membership */
         uint32_t           id;                /**< A copy of `conn->id` */
         mg_addr            addr;              /**< A copy of `conn->peer` */
         int                keep_alive;        /**< Client request contains "Connection: keep-alive" */
